Move business logic from AquariumController to AquariumService

+++ src/main/java/com/example/demo/service/AquariumService.java
@@ -1,13 +1,18 @@
 package com.example.demo.service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import com.example.demo.exception.AquariumNotFoundException;
 import com.example.demo.model.Aquarium;
 import com.example.demo.model.Fish;
 import com.example.demo.model.command.CreateaAquariumCommand;
+import com.example.demo.model.dto.AquariumDto;
 import com.example.demo.repository.AquariumRepository;
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,7 +20,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class AquariumService {
     private final AquariumRepository aquariumRepository;
-    private final FishService fishService;
+    private final ModelMapper modelMapper;
 
     public Aquarium save(final CreateaAquariumCommand aquarium) {
         return aquariumRepository.save(Aquarium.builder()
@@ -30,7 +35,11 @@ public class AquariumService {
 
     @Transactional(readOnly = true)
     public List<Aquarium> getAll() {
-        return aquariumRepository.findAll();
+        List<Aquarium> aquariumList = aquariumRepository.findAll();
+        aquariumList.stream()
+            .map(aquarium -> modelMapper.map(aquarium, AquariumDto.class))
+            .collect(Collectors.toList());
+        return aquariumList;
     }
 
     public Aquarium getOne(final Long id) {
@@ -55,7 +64,14 @@ public class AquariumService {
     }
 
     @Transactional
-    public void remove(final Long id) {
+    public boolean remove(final Long id) {
+        List<Fish> fishes = getOne(id).getFishes();
+
+        if (!fishes.isEmpty()) {
+            return false;
+        }
         aquariumRepository.deleteById(id);
+
+        return true;
     }
 }
